requirements.txt
transformers
tavily-python
openai
